GitHub #146: Tighten up RPC wire protocol and synchronization.
I believe the racy startup stuff was caused by multiple threads accessing the RPC IO
without locking.  This patch fixes that on C and Python side.

Also, I add a magic SYNC token and a record-size so that the wire protocol
receive operation is always for an exact number of bytes.  This gets rid of the need
for kludgy json string manipulations to unpack multiple requests in the same IO
operation.