GitHub #248: Prevent Gtk from grabbing focus
Gtk really likes to change the keyboard focus on TAB or
cursor keys.  We want to keep those keypresses in the Clutter
canvas, so we have to intercept them before Gtk can change
the focus.

Simply taking focus back when it's lost, as we used to do,
somehow locks up the app since Clutter 1.22 was released.